requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.0
django-cleanup==5.2.0
django-crispy-forms==1.11.2
idna==2.10
lxml==4.6.3
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
python-pptx==0.6.19
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
XlsxWriter==1.4.3
